Get tumor types

Get tumor types according to OncoTree's ontology.

Usage

get_tumor_types(oncotree_version = "oncotree_latest_stable")

Arguments

oncotree_version

OncoTree version. Check available options with get_versions().

Value

A tibble of 13 variables:

oncotree_version

OncoTree tumor classification system version.

oncotree_code

Tumor type code: a unique identifier for a tumor type within the classification system of the OncoTree.

oncotree_name

Tumor type name: a brief description of the tumor type.

oncotree_main_type

Tumor main type: a category under which the tumor type can be grouped.

tissue

Tissue associated with the tumor type.

level

OncoTree is a hierachical classification system with 5 levels. At the root level (level 0) there is the single "TISSUE" tumor type. At level 1, there are 32 tissue sites, e.g., "BREAST".

parent

The parent is the parent oncotree_code for this tumor type.

umls_code

The corresponding tumor type identifier(s) in the Uniﬁed Medical Language System (UMLS).

nci_code

The corresponding tumor type identifier(s) in the National Cancer Institute (NCI) Thesaurus.

history

Previous tumor type codes (from previous OncoTree versions) used to identify this tumor type.

revocations

TODO.

precursors

TODO.

color

Color associated with the tumor type.

Examples

get_tumor_types()
#> # A tibble: 885 × 13
#>    oncotree_version    oncotree_code oncotree_name oncotree_main_t… tissue level
#>    <chr>               <chr>         <chr>         <chr>            <chr>  <int>
#>  1 oncotree_latest_st… MMB           Medullomyobl… Embryonal Tumor  CNS/B…     3
#>  2 oncotree_latest_st… GCB           Germinal Cen… Mature B-Cell N… Lymph…     6
#>  3 oncotree_latest_st… SBLU          Splenic B-Ce… Mature B-Cell N… Lymph…     5
#>  4 oncotree_latest_st… OHNCA         Head and Nec… Head and Neck C… Head …     2
#>  5 oncotree_latest_st… PAOS          Parosteal Os… Bone Cancer      Bone       3
#>  6 oncotree_latest_st… TMDS          Therapy-Rela… Leukemia         Myelo…     5
#>  7 oncotree_latest_st… ARMS          Alveolar Rha… Soft Tissue Sar… Soft …     3
#>  8 oncotree_latest_st… SCST          Sex Cord Str… Sex Cord Stroma… Ovary…     2
#>  9 oncotree_latest_st… ITLPDGI       Indolent T-C… Mature T and NK… Lymph…     5
#> 10 oncotree_latest_st… MBC           Metaplastic … Breast Cancer    Breast     2
#> # … with 875 more rows, and 7 more variables: parent <chr>, umls_code <list>,
#> #   nci_code <list>, history <list>, revocations <list>, precursors <list>,
#> #   color <chr>